requirements.txt
$ conda create --name <env> --file 
Keras==2.2.4
numpy==1.15.4
imageio==2.5.0
Flask==1.0.2
tensorboard==1.12.2
tensorflow==1.12.0
tensorflow-estimator==1.13.0
tornado==6.0.2
pillow==5.4.1
